# backend/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: Comfort Curators V0 API Contract Floor
  version: 0.1.0
  description: >-
    Minimum public and operational REST surface for the Comfort Curators V0
    modular monolith. Domain requirements and authorization rules in the
    protected product files remain normative.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/catalog/items:
    get:
      operationId: listCatalogItems
      tags: [Catalog]
      responses:
        '200': {$ref: '#/components/responses/Collection'}
    post:
      operationId: createCatalogItem
      tags: [Catalog]
      requestBody: {$ref: '#/components/requestBodies/Command'}
      responses:
        '201': {$ref: '#/components/responses/Resource'}
  /v1/properties/{property_id}/package-versions:
    post:
      operationId: createPropertyPackageVersion
      tags: [Catalog]
      parameters: [{$ref: '#/components/parameters/PropertyId'}]
      requestBody: {$ref: '#/components/requestBodies/Command'}
      responses:
        '201': {$ref: '#/components/responses/Resource'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    PropertyId: {name: property_id, in: path, required: true, schema: {$ref: '#/components/schemas/Identifier'}}
...
  requestBodies:
    Command:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/CommandEnvelope'}
...
  responses:
...
    Resource:
      description: Resource representation
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Resource'}
    Collection:
      description: Cursor-paginated resource collection
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Collection'}
...
  schemas:
    Identifier:
      type: string
      minLength: 16
      maxLength: 64
      pattern: '^[A-Za-z0-9_-]+$'
    CommandEnvelope:
      type: object
      additionalProperties: false
      required: [idempotency_key, data]
      properties:
        idempotency_key: {type: string, minLength: 8, maxLength: 128}
        reason: {type: string, maxLength: 1000}
        data: {type: object, additionalProperties: true}
...
    Resource:
      type: object
      required: [id, version, data]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        version: {type: integer, minimum: 1}
        data: {type: object, additionalProperties: true}
    Collection:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Resource'}}
        next_cursor: {type: [string, 'null']}
...